Organise
+Organised sources into folders
+Added .bat file for ease of run
+Compatible with VS Code
