+++ assets/testDoc.txt
@@ -1,0 +1,3 @@
+acesta este un text de test
+vai de sufletu meu
+aaaaaaaaaaaaa
