--- testDoc.txt
@@ -1,3 +1,0 @@
-acesta este un text de test
-vai de sufletu meu
-aaaaaaaaaaaaa
